# Conflicts:
#	src/main/resources/templates/main.html

+++ src/main/java/com/trip/dto/RequestPayDto.java
@@ -11,12 +11,12 @@ public class RequestPayDto {
     private String orderUid;
     private String itemName;
     private String buyerName;
-    private Long paymentPrice;
+    private int paymentPrice;
     private String buyerEmail;
     private String buyerAddress;
 
     @Builder
-    public RequestPayDto(String orderUid, String itemName, String buyerName, Long paymentPrice, String buyerEmail, String buyerAddress) {
+    public RequestPayDto(String orderUid, String itemName, String buyerName, int paymentPrice, String buyerEmail, String buyerAddress) {
         this.orderUid = orderUid;
         this.itemName = itemName;
         this.buyerName = buyerName;
